a bunch of stuff - mostly creating pattern 1
-added pot.update to audio update function
-fixed centroid function

note: when pattern1 is made the first pattern, there is a run-time
error, otherwise it runs fine. Although, it's not really tested because
my audio ckt doesn't work right yet.

+++ MSGEQ7.cpp
@@ -112,7 +112,7 @@ byte MSGEQ7::minBand()
 
 float MSGEQ7::centroid()
 {
-	int tmpSum = 0;
+	float tmpSum = 0;
 	
 	for (int n=0;n<7;n++)
 	{
